testing https://github.com/apache/beam/pull/24274 and https://github.com/apache/beam/pull/22179
