+++ canonical-streaming-pipelines/src/main/java/com/google/cloud/pso/beam/pipelines/transforms/WriteFormatToBigQuery.java
@@ -19,6 +19,7 @@ import com.google.api.services.bigquery.model.TableSchema;
 import com.google.cloud.pso.beam.pipelines.options.BigQueryWriteOptions;
 import com.google.cloud.pso.beam.pipelines.options.EventPayloadOptions;
 import java.util.Random;
+import org.apache.avro.generic.GenericRecord;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryHelpers;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryIO;
 import org.apache.beam.sdk.io.gcp.bigquery.BigQueryUtils;
@@ -32,35 +33,41 @@ import org.apache.beam.sdk.values.ValueInSingleWindow;
 
 /**
  * Writes to BigQuery using StorageWrite API.
+ *
+ * @param <T> The type to write into BigQuery, currently GenericRow and Row are supported.
  */
-public class WriteToBigQuery extends PTransform<PCollection<Row>, PDone> {
+public abstract class WriteFormatToBigQuery<T> extends PTransform<PCollection<T>, PDone> {
 
-  WriteToBigQuery() {
+  WriteFormatToBigQuery() {
   }
 
-  public static WriteToBigQuery create() {
-    return new WriteToBigQuery();
+  public static WriteFormatToBigQuery<GenericRecord> writeGenericRecords() {
+    return new WriteGenericRows();
   }
+
+  public static WriteFormatToBigQuery<Row> writeBeamRows() {
+    return new WriteBeamRows();
+  }
+
+  protected abstract BigQueryIO.Write<T> createBigQueryWriter();
 
   @Override
-  public PDone expand(PCollection<Row> input) {
+  public PDone expand(PCollection<T> input) {
     var options = input.getPipeline().getOptions().as(BigQueryWriteOptions.class);
 
     var write
-            = BigQueryIO
-                    .<Row>write()
+            = createBigQueryWriter()
                     .withMethod(options.getBigQueryWriteMethod())
-                    .useBeamSchema()
                     .withoutValidation()
                     .withWriteDisposition(BigQueryIO.Write.WriteDisposition.WRITE_APPEND)
                     .withExtendedErrorInfo();
 
     if (options.getTableDestinationCount() > 1) {
-      var tableSchemaString = retrieveTableSchema(options).toString();
+      var tableSchemaString = BigQueryHelpers.toJsonString(retrieveTableSchema(options));
       var tableCount = options.getTableDestinationCount();
       var tableSpec = options.getOutputTable();
       write = write
-              .to(new DynamicDestinations<Row, Integer>() {
+              .to(new DynamicDestinations<T, Integer>() {
                 private final Random rand = new Random();
 
                 @Override
@@ -100,9 +107,34 @@ public class WriteToBigQuery extends PTransform<PCollection<Row>, PDone> {
     return PDone.in(input.getPipeline());
   }
 
+  public static class WriteBeamRows extends WriteFormatToBigQuery<Row> {
+
+    WriteBeamRows() {
+    }
+
+    @Override
+    protected BigQueryIO.Write<Row> createBigQueryWriter() {
+      return BigQueryIO
+              .<Row>write()
+              .useBeamSchema();
+    }
+  }
+
+  public static class WriteGenericRows extends WriteFormatToBigQuery<GenericRecord> {
+
+    WriteGenericRows() {
+    }
+
+    @Override
+    protected BigQueryIO.Write<GenericRecord> createBigQueryWriter() {
+      return BigQueryIO
+              .writeGenericRecords();
+    }
+  }
+
   private static TableSchema retrieveTableSchema(EventPayloadOptions options) {
     return BigQueryUtils.toTableSchema(
-            PrepareBQIngestion.retrieveRowSchema(options));
+            TransformTransportToFormat.retrieveRowSchema(options));
   }
 
 }
