+++ pom.xml
@@ -14,7 +14,7 @@
 
   <properties>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <beam.version>2.44.0</beam.version>
+    <beam.version>2.46.0-SNAPSHOT</beam.version>
     <java.version>11</java.version>
     <slf4j.version>1.7.36</slf4j.version>
     <maven-compiler-plugin.version>3.10.1</maven-compiler-plugin.version>
